Created Variant Subsystem
Variant subsystem options were added to the actuator and sensor
Simulink files, so users can define which parameters to use for their
specific sim.  The options are currently all the same code and will
need to be edited eventually.

Code was added to the ParamDef file to create the variant choices and
variables.  The added code works with the Simulink models as a stand
alone file, but it does not currently work as part of the ParamDef
file.  Issue is being worked on.