Added back old Ultrastick25e coef
Created a config using the old UltraStick25e aero coefficients and the
new aero model. I removed the -1 on the CYw that was in the old model
and applied -1 to each of the CY coeficients in the config setup. I also
split the aileron and flap coefficients to retain and left side and
right side independent.