Several fixes to attempt to repair aeromodel changes
Fixed aileron and flap control surface coefficients, they were being
halved which was incorrect.  Zeroed CL.alphaDot because it was not
actually used in the old aero model. Changed aileron splitting in the
'trim helper' subsytem to be (ailR = -ail ; ailL = ail) as in the old
block diagram (this will need to be modified again to compare with the
AVL model. Fixed order of coefficients in C6.

Currently spmod, latmod match closely with the original. I believe the
drag model is slightly different and needs further investigation, the
effect can be seen in the throttle trim and increased damping in the
phugoid mode.