Aerospace Blockset Removal and SIL work
1) More Aerospace Blocket was removed from 6DOF block and work was done
to set up a model reference for the environmental model, just as was
done for the navigation model. Aerospace Blockset has been partially
removed from the simplified Environmental model.

2) Lots of work was required to get the SIL back in action. For one, a
bunch of files were moved around and renamed, which broke lots of
things. Items such as buses and trim condition items were changed, but
these changes were not reflected in the SIL.

In particular, the old sensors block was brought back. The new one had
no euler angles, since these are computed in software. however, the sim
does not support this software emultation at this time, thus we need to
keep euler angles in the "sensor" block, or else none of the controllers
work.