Eliminate Aerospace Blockset from Environment Model and Cleanup
Fully functioning environment model working without aersopace blockset.
The only change was a minor delta in the altitude dynamics... which went
from a very low frequency stable pole to a pure integrator in the
linearlized dynamics.

Also cleaned up the setup files for NL and SIL. Several things were
broken in the background, but they work now. Tested the sims with and
without aerospace blockset, and with simulink and c-code controllers.
All looks good!