Sensors using Configurable Subsystems
Variant Models ended up being too much trouble to manage. The basic
problem is that the parameter structures need to be passed as an
argument to the Variant Model. A prototype structure needs to be used as
a seed to the Variant Model when its created. If the structure changes
the model would need to be re-seeded with the new structure, via a small
set of command line commands. However doing this programatically was
painful and managing the system was painful even in a small example
case.

Using configurable subsystems is more straightforward for the sensors
and actuator models. However, the Block Choices need to be handled
programatically and can only be changed with the model open and knowing
the full path to the subsystem.